requirements.txt
python-dotenv
flask
requests
ngrok
wasenderapi
apscheduler
pytz
cryptography